fix(mobile): cast download progress to double
Avoid iOS build failure by converting clamp() result to double for audio download progress.

+++ trmultichat/mobile_flutter/lib/src/core/ui/attachment_preview.dart
@@ -231,7 +231,7 @@ class _AudioPlayerScreenState extends ConsumerState<_AudioPlayerScreen> {
         onReceiveProgress: (rcv, total) {
           if (total <= 0) return;
           final p = (rcv / total).clamp(0, 1);
-          if (mounted) setState(() => _progress = p);
+          if (mounted) setState(() => _progress = p.toDouble());
         },
       );
       await _player.setFilePath(file.path);
